Separate geometry class from scene class
