+++ Renderer/Renderer/RayCasting/SceneGeometry.cpp
@@ -1,0 +1,7 @@
+#include <RendererStdAfx.h>
+
+#include <Renderer/RayCasting/SceneGeometry.h>
+
+namespace GLEngine::Renderer {
+
+}
